Hot fix for CRAN RNG Issue with r-devel (close #15) (#16)
* Add suggested temporary fix by Kurt to R 3.6.0's RNG change.

* Minor documentation spacing fixes.

* Bump version

* Update NEWS with changes.

* Remove integer casting...

* Provide reason for patch..

+++ src/utilities.cpp
@@ -207,13 +207,13 @@ arma::mat sim_eta_matrix(unsigned int K, unsigned int J, const arma::mat &Q)
     return ETA;
 }
 
-//' Simulate all the Latent Attribute Profile \eqn{\mathbf{\alpha}_c} in Matrix
-//form
+//' Simulate all the Latent Attribute Profile \eqn{\mathbf{\alpha}_c} in
+//' Matrix form
 //'
 //' Generate the \eqn{\mathbf{\alpha}_c = (\alpha_{c1}, \ldots, \alpha_{cK})'}
-//' attribute profile matrix for members of class \eqn{c} such that
-//\eqn{\alpha_{ck}} ' is 1 if members of class \eqn{c} possess skill \eqn{k} and
-//zero otherwise.
+//' attribute profile matrix for members of class \eqn{c} such that 
+//' \eqn{\alpha_{ck}} ' is 1 if members of class \eqn{c} possess skill \eqn{k}
+//' and zero otherwise.
 //'
 //' @param K Number of Attributes
 //'
@@ -263,8 +263,8 @@ arma::mat attribute_classes(int K)
 //' Generate a sample from the
 //'  \eqn{\mathbf{\alpha}_c = (\alpha_{c1}, \ldots, \alpha_{cK})'}
 //' attribute profile matrix for members of class \eqn{c} such that
-//\eqn{\alpha_{ck}} ' is 1 if members of class \eqn{c} possess skill \eqn{k} and
-//zero otherwise.
+//' \eqn{\alpha_{ck}} ' is 1 if members of class \eqn{c} possess skill \eqn{k} 
+//' and zero otherwise.
 //'
 //' @param N      Number of Observations
 //' @param K      Number of Skills
@@ -316,7 +316,7 @@ sim_subject_attributes(int N, int K,
         probs_ = Rcpp::as<arma::vec>(probs);
 
         // Verify length is okay.
-        if (probs_.n_elem != (int)nClass) {
+        if (probs_.n_elem != nClass) {
             Rcpp::stop("`probs` must have %s elements instead of %s.", nClass,
                        probs_.size());
         }
